refactor: 요구사항에 따른 코드 리팩토링
GraphService 일급 컬렉션을 활용하도록 코드 리팩토링
PathService 일급 컬렉션을 활용하도록 코드 리팩토링, 메서드 분리 수행
FavoriteResponse 도메인 로직 제거
FavoriteService domain -> dto 변환 로직을 FavoriteResponse에서 옮겨옴
